a lot of core changes

+++ lib/data/models/create_new_main_task_req.dart
@@ -1,7 +1,7 @@
 class CreateNewMainTaskReq {
   String? processingManagerId;
   String? harvestManagerId;
-  int? targetKilosToSale;
+  double? targetKilosToSale;
   int? priority;
   String? destination;
   String? status;
